Revert "Use f32 instead of f64 for armv6 compatibility."
This reverts commit aa5e4ae50d0bf7f726e58faf4caf236586a8b6d4.

+++ src/value/from.rs
@@ -70,7 +70,7 @@ impl From<f64> for Value {
     /// # }
     /// ```
     fn from(f: f64) -> Self {
-        Number::from_f64(f as f32).map_or(Value::Null, Value::Number)
+        Number::from_f64(f).map_or(Value::Null, Value::Number)
     }
 }
